Completed User Line
Closes #25
